Moved content meta preview component into common
